pyproject.toml
[tool.poetry]
name = "blog"
version = "2.0.0"
description = "Refactor"
authors = ["Phenix <eiphenixg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2"
djangorestframework = "^3.13.1"
mysqlclient = "^2.1.0"
django-environ = "^0.8.1"
django-extensions = "^3.1.5"
djangorestframework-simplejwt = "^5.0.0"
drf-yasg = "^1.20.0"
celery = "^5.2.3"
redis = "^4.1.4"
django-redis = "^5.2.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
skip = ["migrations", ".venv"]

[tool.black]
exclude = '''
.venv
| migrations*
| manage.py
| myblog*
| apps.py
'''